feat: Main Train Characteristic Detector
This first commit implements is the baseline of Train Characteristic Detector.

It includes already all the code necesary to compute and plot the rail view  and all the steps before (raw-matrix, filtered-matrix, sobel-matrix, threshold-matrix, mean-filtered-matrix, mask-matrix and rail-view) from a  a given input waterfall.

The function that performs all the pipeline processing is ready to receive as input a numpy array and return a dictionary containing the defined characteristics schema, signal_processor and train_char objects.

The train characteristic schema that this service will return is the following:

    "datetime": None,
    "event": None,
    "status": "not-computed",
    "direction": None,
    "rail-id": None,
    "train-id": None,
    "speed": None,
    "speed-magnitude": "kmh",
    "speed-error": None