refactor: Introduce "train-class" concept
For the moment we don't want to show to the client the "class" of the train ("S-102", "S-103", etc) we had defined until they verified to us that these are the train classes "official" nomenclatures.

Apart from that some semantical refactoring has been made to improve the code understanding.

For this reason, we the train-id will pass now to be an numeric value, like "rail-id", improving in this way the understanding of "id" concept, and "class" will be used as keyword to define the class of the train (Ej. "S-102", "S-103-double", etc.)

For the moment the class will not be given to the client, only a value to remind them that they have to verify to us "train classes"