refactor: Detect unknown event
The unknown events we have detected so far, makes an strong noise in al the railm, where the vibration is able to succed by far the speed limit of 400 km/h (~3750 Km/h)

Ex. 16/03/2023 14:18:16 was detected one of these unknown events.

We think that is a crunch in the rail transmited by the viaduct.