feat: Implement Rail-id Confidence
If the energy difference of each section (power spectral density sum of each section that monitores each rail) is low, two monitored sections are recording similar noise, so, we cannot be sure that we are correctly catching the rail the train is passing throught.

We have detected a 100% correlation between a low rail-id-confidence (similar energy differences between two sections) and the train-class "S-102-6p".

This is the only train class that happens this uncertancy.

We think that this could be happening due to this train enters in resonance with the rail's "natural frequencies".