improvement: Return the best distance confidence
This will improve the accuracy of the clusterization process a lot.

Allows to get each base scores of an specific class being independent from each other.

In this way of proceeding, we can just let as base data the outlayers we had detected. We should be of course, sure that we had tag them properly, so the client should verified to us the class of this specific trains.

It is spected to improve the classification accuracy in the future and having a proper set of base samples, a lot. We could just make the fix and arrive to the 100% of accuracy every time we want, by adding new outlayers as base values.

This is more profound change that it seems.