feat: Automatic serialization of a train track as base data
Stores computed train track as base data.

By default it will be assumed that the computed value is correct.

In the most cases, it is interesting to add to the base data collection a new train track (footprint) when we realize that the computed value is not correct for sure (because it has been verified with the client, adif). For this situation you should tag the data manually using "--class-base" argument option